# Conflicts:
#	FireTreeTrillion/Client/Default/Client.vcxproj
#	FireTreeTrillion/Client/Default/Client.vcxproj.filters
#	FireTreeTrillion/Client/Private/AnimToolHelper.cpp
#	FireTreeTrillion/Engine/Public/Model.h
#	objects_txt/112808_Racing_Kickables.txt
#	objects_txt/112808_Racing_Monsters.txt
#	objects_txt/FinalBoss_Triggers.txt

+++ FireTreeTrillion/Client/Private/LevelChanger.cpp
@@ -48,8 +48,8 @@ CLevelChanger::LEVEL_DATA CLevelChanger::Load()
     json data;
     input_file >> data;
     tLevelData.strObjectName = data["name"];
-    tLevelData.fKirbyHP = stoi(to_string(data["HP"]));
-    tLevelData.fKirbyCoin = stoi(to_string(data["Coin"]));
+    tLevelData.fKirbyHP = stof(to_string(data["HP"]));
+    tLevelData.fKirbyCoin = stof(to_string(data["Coin"]));
 
     input_file.close();
     return tLevelData;
